@@ -24,8 +24,8 @@
   "dependencies": {
     "jotai": "^1.1.3",
     "node-fetch": "^2.6.1",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0"
+    "react": "^18.0.0-alpha-d5de45820-20210714",
+    "react-dom": "^18.0.0-alpha-d5de45820-20210714"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.14.1",
